requirements.txt
ansible-runner==2.3.4
Flask==3.0.1
flask-cors==4.0.0
jinja2==3.1.2
pyyaml==6.0.1
flask-socketio==5.3.6
Flask-SQLAlchemy==3.1.1
flask-login==0.6.3
SQLAlchemy-Utils==0.41.1